+++ http4k-core/src/main/kotlin/org/http4k/websocket/WsClient.kt
@@ -1,0 +1,7 @@
+package org.http4k.websocket
+
+interface WsClient {
+    fun received(): Sequence<WsMessage>
+    fun close(status: WsStatus = WsStatus.NORMAL)
+    fun send(message: WsMessage)
+}